1_Foundations of AI Engineering/012_MLops end-to-end project/insurance -claim-prediction-mlops/requirements.txt
pandas
numpy
ipykernel
mysql-connector-python
sqlalchemy
pyyaml
scikit-learn
imblearn
mlflow

-e .